processing_container/requirements.txt
paho-mqtt
telepot
numpy
opencv-python
